extract hover components
