+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/hover/BenchmarkPill.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import {
   HeatmapHoverBenchmarkPill,
   HeatmapHoverBenchmarkPillProps,
-} from './HeatmapHoverBenchmarkPill';
+} from './BenchmarkPill';
 import {
   BenchmarkComparisonMethod,
   BenchmarkOutcome,
